alarm version 1
comment 작성 시 alarm db 추가
version up 예정

+++ src/main/java/com/example/demo/service/AlarmService.java
@@ -3,7 +3,7 @@ package com.example.demo.service;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
 public interface AlarmService {
-    SseEmitter subscribe(Long id, String lastEventId);
+    SseEmitter subscribe(int id, String lastEventId);
 
     int countReadYn(int userIdx);
 }
